pass startingVerse as a prop to optimize rerenders
